Update To Version 2.0.0 (#3)
* Started working on a new home page for my website.

The current home page isn't good, so I am updating it along with the color scheme.
Also updated some configuration items.

* Finished the skill section.

This took a while, but it looks good. It can have more icons added onto it.

* Added the Contact Page again.

It'll need many more updates.

* Updated the Navigation Bar and the Contact Page.

I am using the Roboto and Lexend Fonts to make this website look far better.
Added the socials section to the Contact Page.
Changed the top bar so that the contact button is where the color button used to be.

* Small edits to some pages.

The mobile form of the contact page looks far better now. No overflowing text and a managed background.
The skills section had a display error that I fixed and now there's better logic.

* Removed all the color handler and route handler logic.

All components and webpages have been updated accordingly.

* Made the start section for the home page.

This replaces the old start section which wasn't that good.
Also made a few edits to the skills section.

* Made The Navigation Home Component.

This component sits at the bottom left and navigates to each of the home sections.

* Grammer edits to Index.js

* Added GitHub Actions to push updates to mohit-jain.com.

* Added a new button to the skills sector.

This takes the to the skills page.
Also added the experience section and modified the home navigation menu accordingly.
Fixed a bug with the mobile routes menu.

* Removed the globe page.

This will save costs, as users will now be redirected to globe.mohit-jain.com in the future.

* Removed unnecessary code.

Removed "Regular" FA Icons and all the globe logic and components.

* Organized some styles.

Removed the globe styles.
Put the Navigation card styles in the actual component itself.
Adjusted other components necessarily.

* fixed build issue

* changed config to not include Map Key.

* Organized some more styles.

Removed many of the CSS files and instead put their styles directly in Vue.js components themselves.
Moved the Route Menu directly into the Navigation Main Panel.

* Organized some styles for the update pages.

* Made most of the experience section.

The last part is creating the phone that sits on the right.
Also using pages the github website for the updates page to reduce costs.

* Finished making the experience section for the main page.

Now the phone on the right is complete.
There's a new Date Store that gets the time from online.
Removed the old page sectors entirely.
added the sublo website link.

* Made The Update 1.4.0 Page.

This has info on the new homepage, new contact page, and new domain.
The Updates page will now always display updates for the most recent update page.
Navigation Home now sits at the top left and its icon beats for 5 seconds.

* Organized More styles.

Now the info sector styles are in their proper places.

* Config Changes.

Hopefully this works...

* Trying deployment again

* Update develop.yml

* changing config

* undoing my stupid commits

* Create gh-pages.yml

* Update gh-pages.yml

* Just removed all the changes i made.

* Changed the "old" website link

* Create gh-pages.yml

* Just reverting stuff again. I acted too hastily with no end results.

* Made a new, far more improved version of my resume.

also removed the static.yml file.

* Initialized the Credits Page.

This page will give credit to certain libraries and services I use while making this website.

* Finished the Credits page.

Simple, but in a good way.
Adjusted the Home page as well to make a path to the credits page.

* Made more sections to the credits page.

Changed the Cesium Icon to an official SVG and adjusted other components accordingly.
Made slight changes to the AWS page descriptions.

* Updated The info sectors.

Added an info sector for Amazon WorkMail.
Adjusted the Main iVue info sector.
Adjusted other components to accommodate these changes.

* Added a Copyright Section at the bottom of the home page.

Added a credits option on the Navigation Bar.
Added two cards to the Credits Page.

* Made an Alert Box for the contact page.

This informs the user of whether they copied a link for their use, and it also informs the user that they sent an email using the API.
Changed the background of the credits page.

* Changed the 1.4.0 update page to the 2.0.0 update page.

Also added an updated resume.

* Edited the Version number for the vue.js project.