Update 2.1.0 (#6)
* Updated the Contact Page.

This now has some conditions for making a message.
Also edited some design issues with some of the homepage's components.

* Made lots of changes to the skills nav page.

* Made the credits section the new skills section.

now there are more cards for all the skills i know. I am not finished with the descriptions though.

* Fixed up all the descriptions for the skills section.

* Started up some new stuff.

Installed animate.css for my homepage.
Made a initial timeline page. It's still in its infancy.

* Added unique transitions.

This is thanks to Animate.css and vue-observe-visibility, which are two libraries that easily let me add transitions.
Made some other styling edits as well.

* Made the home footer its own component.

Now it sits on the home page and the contact page quite well.

* Updated Card Pages.

Added an "Expo" and "More Info" card to the skills section on the homepage.
Added the footer and slightly changed the transition for the skills and experience nav card pages.

* Focused On Icons.

On the footer, I made an option for people to see my Icons.
Also Made the Gold Icons, which can be used for a future project of mine.

* Trying something new out with meta tags for a child path.

* Moved all the icons to the static-icons folder.

Also gave it its own index.html folder that redirects the user to the icon page.

* Fixed build issue with my website footer.

* Made the start section for the icons page.

Changed the main route to be the icons route.

* Made the InitPage Component.

This component, when used, initializes a page.
I also installed @unhead/vue for my project for Search Engine Optimization (SEO).

* Added transitions to some components of the homepage.

Also used the Init page component with more pages.

* Removed the old timeline page.

Also changed some imports to use the @ symbol.

* Added two new skill entities for Vite and Nuxt.

* Updated the Contact Page Social tabs.

Now they have more color.
Also changed the tablet screen width to be 825px instead of 800px.

* Finished Up My Updates for 2.1.0

Skill Widgets are online.
Removed the Update section entirely. I'll just make it again with VitePress, VuePress, or something else.