Version Update 2.4.0 (#13)
* Made the skill note component.

This component is great for a new skills page rework.

* Made new Nav Cards specifically for the home page.

These will lead to certain sections on the skills page given time.

* Removed many pages.

These pages are now being removed or  reorganized to make my website easier to navigate.

* Removed unnecessary components.

* Multiple different changes primarily to Skills Home Section.

Reorganized the skill notes so that the more info note resides in the skills section component.
Changed some styles and text in the skills section to make it more appealing.
Made navigation with the web footer and home navigation widget slightly better.
changed the animation on the exp page.

* Emptied the Skills Page.

The skills page will be reworked into something far better.
Also deleted some more unnecessary files.

* Skill Section Updates + npm package update.

Many core npm packages have been updated, along with the project version number.
The skill note has a transition, includes the "More Info" note, and accommodates for any native website path.
Removed a few more unnecessary particleTS objects and note.css.

* Reduced the number of routes and pages.

This makes the built vue.js + vite-ssg project lighter, allowing for lower latency and costs.
Updated sitemap.xml.

* Started adding a body to the skills page.

Nearly all the layout is complete. The main thing that's missing is the descriptions.

* Update SkillsNav.vue

* Made the rest of the layout.

Now the descriptions are all that's left.
Doesn't look the best on a mobile device, but it'll work for this update.

* Completed Most of the Skill Notes for the Skills Page.

There are many AWS skill notes remaining to be finished, but everything else is completed.

* Completed the new AWS Section for my website.