Version Update 2.5.0 (#16)
* Installed Unplugin-Vue-Components.

This module lets me auto import any component, including the Navigation Bar, the Footer, and any other components.
Also changed the Web Footer to work better.

* Installed Unplugin-Auto-Import.

Similar to the previous module, this module makes it easy to automatically import stores and objects.

* Fixed Build issue on QR Code Page.

* Initialized the projects page.

This page will show some of the projects I worked on.
In terms of layout, it'll use an extremely similar design that the experience page has.

* Completely Changed The Navigation Bar.

Now it floats at the top of the screen and looks better.
Other Pages Now accommodate for this new navigation bar.

* Fixed some bugs regarding closing the top bar.

Now it closes whenever the visitor scrolls on a page or touches the page via any means.
When clicking on the image on the top bar, it scrolls up.

* Linked the Mapping Neurodivergence hackathon project to my website.

I made a project card for the project.
I added assets for the project card.
A link here will redirect the user to the web app.

* Created the Globe Nav Card.

Also made a universal function for the mounting of the navigation cards' initial animations.

* Fixed build issue.

* Made the pizza project card.

This Pizza Project card leads to the mock UI for a simple project of mine that I made Fall 2024.
It also has a route tied to the project.

* Update Cache Id.