Version Update 2.6.0 (#19)
* Set Up the Gamepad Store and Installed the Gamepad Wrapper.

Also changed the PWA cache ID in advance.

* Using Joypad.js for working the Gamepad API.

Also edited and removed some unnecessary stuff.

* Got some small functionality with the controller now.

* Moved the Navigation Menu Component to App.vue

This is so every page has the Navigation Menu and doesn't exist separately on each page.
Also made the navigation menu open on some controller inputs.

* Updated Routing overall.

Clicking on a bumper or a trigger will move onto the next or previous page depending on the app.
The "Invalid Route" Page doesn't rely on a component anymore.
The Navigation Bar has less Icons at the top for navigation.

* Made a Bookmark Component.

This updates as the page changes to make it easier to tell what page the visitor is on.

* slight edit with skills page.

* Started working on the custom cursor.

Like a regular cursor, this will be able to click on buttons and move around.

* Added all Joystick capabilities.

The left stick controls a custom cursor for the website.
The right stick can control scrolling.
The Bookmar cmponent has also be updated to account for more pages and hosts a Gamepad bookmark on the right of the screen.

* Handled Multiple Errors.

Perviously, only "a" elements were clickable by the custom cursor. Now, regular buttons are clickable as well.
Before enabling the Gamepad API, the website checks if the Gamepad API is available before enabling it.
The Icons page now has its own popup that can handle errors.

* Made Multiple Changes.

Added the screen wake lock to the website. If active, it will keep the computer on past when it automatically stops.
Rewrote the logic for the custom cursor position so it uses less variables and fixed some bugs.
Edited routing so that it looks better.

* Set alert for screen wake lock.

* Updated The Bookmarks.

I moved the Gamepad Bookmark to the bottom left.
The new Screen Wake Lock Bookmark is visible when the user enables the screen lock.
The Navigation Bar accommodates for the top bookmarks no matter the screen size now.

* Made the gamepad work a lot better.

The axis threshold has been lowered for more sensitivity, making the joystick work a lot better.
The Directional Pad also controls the cursor.
The right joystick controls scrolling better now.

* Set a custom port for vite preview.

* Added the Repository Link to the website.

This will simply lead vistiors to the repsoitory of my website.
Also updated the copyright statement.

* Added vite-ssg-sitemap

This dev dependencyautomatically generates the sitemap.xml file and robots.txt file upon running npm run build.