Version Update 2.6.2 (#21)
* Updated Document Routing.

Now there is a pdf version of the certificate page on the website.
Certificate page has a new route.
Updated JavaScript logic to accommodate for these changes.

* Updated dependencies + Added Gamepad Page.

Now all installed dependencies are up to date.
There is also an empty page for the gamepad now.

* Made more paths to navigate to the Gamepad Page.

Now both the footer and the Navigation Menu has buttons leading to the page.
Also made some other edits to the Navigation Menu.
Gave the Gamepad Menu its initial layout.

* Removed the gamepad page.

Using the gamepad is straightforward enough.
I made some other light edits as well.

* Added Sigificantly More Gamepad Functionality.

When hovering over a button, the Custom Cursor will change to a pointer Icon.
The App now also tracks if gamepads are connected or not and displays them at the bottom right.