Version Update 2.6.3 (#22)
* Improved ergonomics of the custom cursor.

It naturally moves faster now.
When holding down X or Y, the cursor halfs its overall speed.

* Added unplugin-swc.

This reduces the build time for my website by over a second.
Also updated the project version and the cache ID for my website.

* Created the Documents Section.

The documents section simply guides the user to one of my documents. It looks identical to the iVue Section.

* Small Changes.

Added an icon on the FCS Certificate page to lead visitors to the Fulton schools internship program.
Swapped the positions of the lock bookmark and the gamepad bookmark.
Made sure that the cursor changes as the window scrolls or resizes.

* Added animation to custom cursor.

It pulses when it hovers over a button or link.
Also fixed a flaw in the website footer.