Version Update 2.6.4 (#23)
* Added vite-plugin-qrcode + updated dependencies.

this new plugin shows the network connections as qr codes so it's easy to view the app on phones.
Some dependencies needed to be updated, so i went ahead and did that too.

* Added text to speech on contact page.

Now the message can play out if the user wants to hear it.

* Added Gamepad Vibrations for clicking on a button or link.

Doesn' trigger with external links, but works with everything else.