Version Update 2.6.6 (#24)
* Added Unplugin-Info and Updated Copyright Page.

This npm module provides information about the last release of my website and other great information that can be displayed to the users.
The copyright page has been updated.

* Added a safeguard to the screen wake lock.

The website now checks to ensure that the WakeLock API exists upon opening the app. If it doesn't, the icon to set the wake lock indicates that.
Also changed the download icon for the resume and certificate pages.

* Made the Install Page.

This page will simply host a button prompting the user to install the website. Added link buttons on the Footer and Navigation menu that lead to this page.
Made a few template edits to the Navigation Bar.

* Completed the Install Page.

Made a new Pinia Store to handle the logic.

* Forgot the click hook