Version Update 2.7.0 (#28)
* Website Now uses more SEO friendly tags.

Also added the pagefind UI for searching on my website.

* Added the search page to the website.

Made the navigation bar and footer clinet only components and had them accommodate the new page.
Feature Not completed yet.

* Patching build issue. will ruly handle later.

* Fixed UI Issues on the contact page.

Now the footer should look correct again.
Also replaced Pagefind with Fuse.js because it is more customizable and better for my use cases.

* Added querying on the search page.

* Updated the iVue Experience Card with more accurate knowledge.

I updated the link for the Main iVue Website and everything the card says.

* Organized the JavaScript for the Search Page.

* Updated the Search Index for the search page.

It now includes all the website's pages.

* Added the Click Sound effect.

Whenever the user clicks on a link or a button, a sound effect plays.
Also removed unnecessary classes and changed the custom cursor class into an ID.

* Added a menu that is used for controlling the cursor speed.

It uses the bumpers right now, but it can use more features in the future.
Made minor improvements in other places as well.

* Added the gamepad page.

Empty right now, but this will show what controls are used for a gamepad at any time.

* Made the Gamepad Controls page.

This just lists the controls now for using a gamepad with my website.

* Updated dependencies