Version Update 2.7.1 (#29)
* Made the Gamepad page in the form of a popup instead of a page.

This makes it a bit easier to manage.
The Navigation Bar nor the website footer is on this page anymore. It simply has a button that takes the user back to the previous page.

* Improved searching on the search page.

Rather than the app performing a search every time a new letter is entered, a new button and the enter key now do it instead.
It is also called if the user enters something but stops for at least 2 seconds.

* Updated Vite to version 7.0.0.

Initially, other packages were unable to handle vite 7.0.0, but now they are capable of it without any major problems.

* Created .node-version for Cloudflare pages testing.