Version Update 2.7.4 (#32)
* Added a rounded icon.

The icon page should now display this new icon and the MND Icon.
I updated the meta tags and the PWA to use this icon instead of the square.

* Globe Navigation Card Updates + Updated Dependencies.

Updated the Globe Navigation Card and other related components and assets to properly reflect the new features present on my Globe UI and its purpose.
Updated the Vite and Unhead dependencies. Also updated the project version number.
Made other small edits on the iVue Nav Card and the create app function on main-ssg.js.

* Add Tauri Skill Note + Added Version Number To Copyright Page.