Removed the Search Page and made Config updates.
The serach page is quite useless, so I removed or edited anything associated with it. It'd be easier for me to make search functionality another way.
I updated the Project Version Number and PWA Cache ID to v2.8.0
I updated umplugin-auto-import to Version 28.
Changed some stuff to make it easier to reach the contact page.