Added significantly more logic to the click audio. See description.
Now the audio volume is stored in the website's local storage so that it stays changed even when the user closes out of the website.
Also made it possible for a gamepad to edit the volume with its triggers.
I moved all the code for this feature to its own Audio Store.