Version Update 2.8.0 (#36)
* Removed the Search Page and made Config updates.

The serach page is quite useless, so I removed or edited anything associated with it. It'd be easier for me to make search functionality another way.
I updated the Project Version Number and PWA Cache ID to v2.8.0
I updated umplugin-auto-import to Version 28.
Changed some stuff to make it easier to reach the contact page.

* A Couple Changes. See Description.

Put links on the projects section and the skill section headers.
Made the "mohit-website" link and any subroute redirect back to the main route.

* Small Text Edits

* Added a new tab for Florida Man.

This is a new production by iVue Media. Thought it should be included.

* I made a volume meter for the click sound.

This is on the Nav Menu.

* Added significantly more logic to the click audio. See description.

Now the audio volume is stored in the website's local storage so that it stays changed even when the user closes out of the website.
Also made it possible for a gamepad to edit the volume with its triggers.
I moved all the code for this feature to its own Audio Store.

* Added a dependency to track env types.

Also updated unplugin-auto-import.