Made multiple gamepad changes. See Description.
I made a popup at the bottom left that works exactly like a real cursor hovering over an Anchor element, but instead for the gamepad cursor.
Clicking on any of the 4 right hand buttons emits a click.
Removed the simple-kyeboard npm package. I don't want to implement it right away.