Version Update 2.8.1 (#37)
* Multiple small changes. See Description.

I updated the Vite dependency AND added a new simple-keyboard dependency.
Updated Project Version Number and PWA Cache ID.
I also moved the contact page styles to its own file as it was the largest section on ContactPage.css.

* The custom cursor now accounts for input bars.

When the gamepad cursor is now over an input bar, it will now run the "focus" function on said input bar.

* Made multiple gamepad changes. See Description.

I made a popup at the bottom left that works exactly like a real cursor hovering over an Anchor element, but instead for the gamepad cursor.
Clicking on any of the 4 right hand buttons emits a click.
Removed the simple-kyeboard npm package. I don't want to implement it right away.