Version Update 2.8.2 (#38)
* Font Awesome Icons now render on the server directly rather than the client.

* Updated configuration. See Description.

Updated Project Version Number + PWA Cache ID.
Added the "simple-keyboard" npm dependency back to the website.
Moved all the .d.ts files to the "dts" folder and added it to the .gitignore file.

* Added unplugin-imagemin.

Now images are compressed with Squoosh when the app is being built, which halves the size of the "dist" folder.
Also removed simple-keyboard (again) because it wouldn't be too useful with the website.