Version Update 2.9.0 - Documents (#41)
* Started working on making new pages to display PDF files.

* Got the new module fully working on the website.

Made a lot of other edits too.

* Finished developing all the logic for the new pages.

Now the pages properly resize and all the logic is on the document store.
Also moved the Text To Speech functionality to the Audio store.
I adjusted many otehr files to accommodate these changes.

* COMPLETELY reworked document navigation.

Now the document pages use a navigation bar at the bottom rather than the widgets.
There is a new function that allows vistors to print the document.

* Fixed bug with click event.

* Prevented unnecessary messages from being logged.