Version Update 2.9.1
Updated My resume significantly. Now the pdf file and the markdown component has more elements and whatnot.
Also added "markdown-it-attrs" to add custom classes to some of the elements in markdown files.
Updated the Project Version Number and the PWA Cache ID to v2.9.1. Also updated Vue.js to v3.5.20.