Made multiple edits to the website. See Description.
I made a button for the website features on the projects section of the main page.
Made a new widget for the features pages that return the user back to the original page.
Removed the Icons page from the list of "feature" pages. Adjusted the navigation bar and the footer to accommodate this changes.