Made Updates to naivgation components.
The Main Nav Bar always closes when someone clicks on the "footer scroll" button. Furthermore, it should always scroll now.
The document nav bar now scrolls to the top of the page if the user clicks on the button designated for that page.
The tato pdf components and the markdown components now have a container which ensures that the footer remains below the initial viewport.