Made more functionality for the code-reader.
Upon scanning a new value, a new sound effect is triggered. This is already configured in the audio store and for the PWA.
The trash icon now deletes all scanned items from the list.
I made more of the UI for the scanner so that everything looks cleaner and more functionality can be added.
I made the "truncate" function a global function so that multiple pages and components can use it.