Version Update 3.0.0 (#43)
* Started working on a page that can scan barcodes and qr codes.

I setup the page and installed a dependency.
Also updated Project Version Number and PWA Cache ID.

* Completely made a new features page for the website.

This simply shows off the unique features of the website like the PWA, the qrcode page, etc.
There are also proper buttons that lead the user to this page.
Also made a lot of other edits.

* Reorganized Feature files and hopefully fixed build issues.

* Made multiple edits to the website. See Description.

I made a button for the website features on the projects section of the main page.
Made a new widget for the features pages that return the user back to the original page.
Removed the Icons page from the list of "feature" pages. Adjusted the navigation bar and the footer to accommodate this changes.

* Multiple Updates. See Description.

Added keybinds that let the user navigate to the the wake lock page, the install page, and the code reader page.
I made a button in the navigation menu that takes visitors to the footer. It doesn't appear on pages without the footer.

* Edited some stuff about the footer.

Now the Web Footer component itself states whether it is present or not.
The footer has been added properly to the Main resume page, the Markdown resume page, and the main FCS certificate page.
Made a link to PDF.js considering I use it a lot.

* Multiple updates. See Description.

Started working on the Barcode reader page.
Gave the Features Return Widget a better hover effect.
Added the documents note to the Features Page.
Fixed a few errors.

* Started making more elements for the Code Reader Page.

* Updated Resume + More updates to Code Scanner.

My resume now includes a QR Code at the top that leads users to my website.
The code scanner now has better comments and updates the User Interface with each code scanned on the console.

* Made new version number to be 3.0.0

* Made Updates to naivgation components.

The Main Nav Bar always closes when someone clicks on the "footer scroll" button. Furthermore, it should always scroll now.
The document nav bar now scrolls to the top of the page if the user clicks on the button designated for that page.
The tato pdf components and the markdown components now have a container which ensures that the footer remains below the initial viewport.

* Made more functionality for the code-reader.

Upon scanning a new value, a new sound effect is triggered. This is already configured in the audio store and for the PWA.
The trash icon now deletes all scanned items from the list.
I made more of the UI for the scanner so that everything looks cleaner and more functionality can be added.
I made the "truncate" function a global function so that multiple pages and components can use it.

* Continued making more updates to The Code Scanner Page.

All logic is now stored within a Pinia Store for more convenience.
Scanned items are now saved in the local storage for the website and updated as items are added and removed.
Links that lead to external websites now can be clicked on.

* Made multiple updates. See Description.

Changed the path of the Code reader page to "/code-scanner".
Made the proper logic to open a menu for each scanned item on the code scanner page.
The app is not using the legacy worker from PDF.js. If necessary, I can enable the app to do so again.

* Updated Dependencies + minor code formatting.

* Updated my resume.

This Time I used an AI checker to see how good this would look to be.

* Started making the item menu.

Main buttons are not there yet, but they will be soon.
Also updated the font awesome icon packages.

* Finished the code scanner page.

All the code has been made for this page now.

* Added arduino to this website and my resume.

* Added a new navigation skills widget.

This allows better navigation on the skills page.
I also added the "R Project" skill note.

* Edited the Icons page.

Now instead of copying the image, clicking the button will direct the sure to a new tab with the image.
Made other minor edits as well.