Version Update 3.0.1 (#44)
* Replaced vue3-observe-visibility with motion-v.

Motion-v is a dependency that can provide more functionality than the former while still being friendly for SSG.
I applied this to the homepage for the time being.
Updated a fontsource dependency, the project version number, and PWA Cache Id.

* Incorporated Motion.div into the skills and features pages.

Also Added a "website features" button to the projects page.

* Updated Version number and checked build before deployment.