Version Update 3.0.2 (#45)
* Replaced Joypad.js with native code.

Now all events from a gamepad are made without Joypad.js and work exactly the same. The events now have moderate type support as well.
Uninstalled Joypad.js

* Added a new "button hold" event.

This tracks when a button is being held down and how long it is held down.
I have replaced some functionality when adjusting the max cursor speed and the volume so that it uses the new event rather than intervals.

* Updated a resume link.

On the top, I edited my website link to use www.mohit-jain.com instead of mohit-jain.com.
When a user clicks on it on the main resume page, it will guide the user back to the home page rather than to a separate tab.