Added a new QR Code tool for the website.
This tool allows a QR Code to be available for any page. It can be activated on the Navigation Menu or with CTRL + q.
The "HomeRedirect" page is made specifically for any page that is designated to go back to the home menu.
I made other small edits as well.