Version Update 1.3.0 (#46)
* Added the Full screen store and updated dependencies.

This store will allow the website to set any element to full screen.
I updated both the Font Awesome Vue component and Vite, along with both the project version number and PWA cache ID.

* Implemented the full screen store on the main resume page.

This includes adding a widget, new styles, and new functions and objects.

* Made many edits and rearrangements. See Description.

Moved all the widgets to their own widgets folder.
Added the shrink widget to all the document pages.
Edited many css classes to have all the resume containers look good in full screen mode.

* Added a new QR Code tool for the website.

This tool allows a QR Code to be available for any page. It can be activated on the Navigation Menu or with CTRL + q.
The "HomeRedirect" page is made specifically for any page that is designated to go back to the home menu.
I made other small edits as well.

* Implemented the Qrcode Drop Zone component.

It detcts images as the hover above the component and can be scanned easily. It has been easily integrated with the rest of the page.

* testing new service worker notifications.

* Made many changes. See Description.

Not implementing the Notifications API due to some errors.
Removed uqr and the QR code page. I will implement a better npm package for this.

* Implemented the new QR Code npm dependency.

* Made the final updates before pushing code.

Made a button on the QR Code popup that allows users to download the QR Code.
Added a qr code button at thefooter and another on the features page that makes it simple to open the qr code popup.
I made a few styling edits across the website.