Made MANY updates. See Description.
Finished up my version of the Google Mockup. It has better buttons for both the dark mode and the directions, actually searches stuff on google, and the directions still has a hash link.
Updated some Fontsource dependencies, the project version number, and the PWA Cache ID.
Made routes on both the footer and the features page that leads to the Google Mockup.
Changed the main redirect route to my repository to "repository" and made a card on the features page for tsParticles.