Implemented the Web Share API into the website.
Using the VueUse utility, this websit allows people to share a link or a file with anyone else. I have put this within the Contact Page and the Share popup.
I also use the FPS utility from VueUse to let Gamepad Button hold events track how long a gamepad button has been held.
Also changed the designs for buttons in the Contact page and those associated with the Share popup.