Reworked the documents on my website.
I removed both the Google and the Browser viewers for both, which led to many functions being removed.
I moved the remaining files to the documents folder in the pages folder.
The Document Navigation Barlooks different, as it now has more options as smaller buttons.
Made a function that lets users share a document.