Version Update 3.2.0 (#49)
* Made many updates with the dependencies. See description.

Updated a few dependencies, added Vue use, and removed motion-v.
Integrated useIntersectionObserver into the website in place of motion-v.
Made many other small edits.
Updated the version number for the website.

* Implemented the Web Share API into the website.

Using the VueUse utility, this websit allows people to share a link or a file with anyone else. I have put this within the Contact Page and the Share popup.
I also use the FPS utility from VueUse to let Gamepad Button hold events track how long a gamepad button has been held.
Also changed the designs for buttons in the Contact page and those associated with the Share popup.

* Reworked the documents on my website.

I removed both the Google and the Browser viewers for both, which led to many functions being removed.
I moved the remaining files to the documents folder in the pages folder.
The Document Navigation Barlooks different, as it now has more options as smaller buttons.
Made a function that lets users share a document.

* Added animations to some of the main document options.

* Implemented a function that adds a QR Code at the top right of any Resume I make.

More buttons are soon to come.

* Made the UI for my resume with the QR Code.

Now it has a route, all the buttons work, and everything looks as expected.
Reworked some logic for all documents for downloading, printing, sharing, and identifying their website route.
Made a type declaration for the object returned by usePDF.