Version Update 3.2.1 (#50)
* Made multiple small changes. See description.

Added the VueUse wake lock utility to replace some functionality.
When visitors click on the resume or qrcode icon on the document navigation bar, it scrolls the page up to the top.
Updated unplugin-vue-components, the project version number, and the PWA cache ID.

* Made a url for the blob that represents my QR Code Resume.

Also made an good route that redirects users to that resume in the viewer.

* Made a few edits to the contact page.

When the user starts typing without focusing on an input bar, it types on the Title input bar.
Added a button that opens the share popup.