Version Update 3.2.2 (#51)
* Updated Vite and version number and made new Keybind.

Vite has been updated to v7.1.7.
The version number and cache id has been updated to v3.2.2.
The Keybind Ctrl + Alt + W now toggles the screen wake lock.

* Made a component for the Particles background dependency.

Now, it handles the JS for a background. When the webpage is not visible on the visitor's screen, it pauses the animation, which should help with a laptop's performance.
Added the VueUse document visibility utility to make this new functionality.

* Added the new Particles component to the rest of the website.

* The Qrcode resume now has a route tied to the query rather than a hash.

This now works better with hashes like the footer hash.
Adjusted the website to accommodate this change. That too using queries make more sense.
Also made other minor edits.