Version Update 3.2.5 (#54)
* Made slight edits. See Description.

Updated project version number and PWA cache ID.
Fixed code formatting for the "cursorElementTitle" computed string.
The QR Code Tool now only appears when the app can use client only components and uses the VueUse useScrollLock utility over an event listener.

* Updated the "printDoc" function for the documents.

It should now work better on all devices.

* Got the share popup to use external links now.

It is tied to a query variable to make it accessible across the website.
Currently integrated with the Contact Icons.
Also properly implemented the "useScrollLock" utility with the popup.

* Made more buttons for the qr code popup.

There is a proper copy button and a button to open its link in a new tab.
The popup also has an integer to set its mode.
The Contact Buttons on the home page use this popup now.