Version Update 3.2.6 (#55)
* Updated dependencies and upgraded the footer scroll button.

The version number has been updated.
When the user clicks the footer scroll button while the footer is visible, it takes the user back to the top of the page.

* Made the website only accept inputs from one gamepad at a time.

This is because the app is currently only designed for one controller and doesn't need input from more than one.
Updated the usePDFObject interface so it is a global interface.

* Made the Nav Menu aware of if it was overflowing or not.

This helps with a particular style change for the navigation bar.
Also gave the Invalid page its own tsParticles Background.

* Fixed up some heights so that the footer is always not visible unless the user scrolls.