Version Update 3.3.0 (#59)
* Updated dependencies and project version number.

Also made the auto scroll for the documents section of the home page less intense.

* Fully created a dropdown for the starting social buttons on the home page.

This uses a few VueUse utilities and other existing functionality present in the application.

* Started updating the pdf document viewer.

Now, there are link buttons made for every document so that visitors can copy the links of specific pages on the website.
The website keeps track of when a pdf is fully loaded on the website viewer.
I made other small changes to enhance the website.

* Made a new add ons component for the document viewer.

This contains the link button and a loading cover for when the document its not fully loaded.

* Made a new update box component.

This update box allows the user to update the version of my website when there is a new version in the future.
Widgets now sit on the bottom left corner to leave room for the update box.

* Update vite.config.js

* Small changes.

* Update main-ssg.js

* Update vite.config.js

* Trying some more confinguration.

* Update vite.config.js

* Made the Copyright page display more information.

* Update index.html

* Made a brand new deploy script.

This replaces what I was using on GitHub actions. It is more unopiniated, but now i can set custom Content cache headers for certain files when deploying them to AWS S3.
I imported many new dependencies from AWS and elsewhere to help me.
Integrated new environment variables to the website as well.

* Made a route that redirects to the deploy script.

Also added comments to the script for better readability.