Version Update 3.3.1 (#60)
* Updated Dependencies and project version number.

Also added the NPM install command to the deploy script.

* Updated the deploy script.

It can now take in the folder name for what build output the user would like to use, defaulting to 'dist'.
It also deletes ALL current files in the bucket before pushing new ones in.
Arguments can now be passed in to disable the deletion and the cache invalidation.

* Made a button on the documents section for the deploy script.

Made the link for it on GitHub a global string.
Made the scripts folder have an alias.